Move template implementation to a separate package
